requirements.txt
faiss-cpu==1.8.0
sentence-transformers==3.0.1
transformers==4.41.2
datasets==2.19.1
beir==2.0.0
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
tqdm>=4.66
